Refactored getUtxos() into Blockchain library
