+++ lib/blockchain.js
@@ -28,8 +28,21 @@ class Blockchain {
 
       return balanceResult.balance.confirmed + balanceResult.balance.unconfirmed
     } catch (e) {
-      console.log(e)
+      // console.log(e)
       throw new Error(`Could not get balance for ${cashAddr}`)
+    }
+  }
+
+  // Get's all UTXOs associated with an address.
+  async getUtxos (cashAddr) {
+    try {
+      const utxoResponse = await this.bchjs.Electrumx.utxo(cashAddr)
+
+      return utxoResponse.utxos
+    } catch (e) {
+      throw new Error(
+        `Could not get UTXOs for ${cashAddr}, details: ${e.message}`
+      )
     }
   }
 }
