--- test/integration/util-integration.js
@@ -1,23 +1,0 @@
-/*
-  Integration tests for the util.js utility library.
-*/
-
-// npm libraries
-const chai = require('chai')
-// const { mockSingleSweepWithBch } = require('../unit/mocks/util-mocks')
-// const BCHJS = require('@psf/bch-js')
-// const bchjs = new BCHJS()
-// Locally global variables.
-const assert = chai.assert
-
-// Unit under test
-// const UtilLib = require('../../lib/util')
-
-describe('#util.js', () => {
-  describe('#buildSweepSingleTokenWithBchFromPaper', () => {
-    it('should do something', async () => {
-      // Assert that the hext strings match.
-      assert.equal(true, true)
-    })
-  })
-})
